+++ d2rloader/core/platform_linux/process.py
@@ -1,22 +1,20 @@
-import subprocess
 from pathlib import Path
 from time import sleep
 from typing import TYPE_CHECKING
-
-from d2rloader.constants import WINDOW_TITLE_FORMAT
-from d2rloader.core.platform_linux.utils import get_window_by_title, set_window_title
-
-if TYPE_CHECKING:
-    from d2rloader.core.state import D2RLoaderState
 
 import psutil
 from loguru import logger
 from PySide6.QtCore import QObject, QThreadPool, Signal
 
+from d2rloader.constants import WINDOW_TITLE_FORMAT
 from d2rloader.core.exception import ProcessingError
 from d2rloader.core.platform_linux.umu import UmuManager
+from d2rloader.core.platform_linux.utils import get_window_by_title, set_window_title
 from d2rloader.core.worker import Worker
-from d2rloader.models.account import Account, AuthMethod
+from d2rloader.models.account import Account
+
+if TYPE_CHECKING:
+    from d2rloader.core.state import D2RLoaderState
 
 
 class ProcessManager(QObject):
@@ -26,6 +24,7 @@ class ProcessManager(QObject):
     def __init__(self, parent: QObject, appstate: "D2RLoaderState") -> None:
         super().__init__()
         self._state: "D2RLoaderState" = appstate
+        self.umu_manager: UmuManager = UmuManager(self._state)
         self.threadpool = QThreadPool()
         self.threadpool: QThreadPool = QThreadPool()
 
@@ -76,49 +75,17 @@ class ProcessManager(QObject):
         self.process_finished.emit(result[0], result[1], result[2])
 
     def _start_instance(self, account: Account):
-        umu_manager = UmuManager(self._state.settings.data, account)
-        game_settings = self._state.game_settings.get_game_settings(account)
-        self._validate_start(account)
-
-        if umu_manager.save_start_script():
-            try:
-                with open(umu_manager.start_script_log_path, "w") as logfile:
-                    game_settings.set_account_game_settings()
-                    logger.info(
-                        f"Launching instance: {umu_manager.start_script_path.absolute()}"
-                    )
-                    logger.debug(f"Using GAME_PATH: {game_settings.settings.game_path}")
-                    proc = subprocess.Popen(
-                        ["sh", umu_manager.start_script_path], stderr=logfile
-                    )
-
-                    if self._is_d2r_started(proc.pid):
-                        self._rename_window_title(proc.pid, account)
-
-                    return None, account, proc.pid
-            except Exception as e:
-                logger.error(e)
-                raise ProcessingError(
-                    f"Error occured during executing {umu_manager.start_script_path}", e
-                )
-
-    def _validate_start(self, account: Account):
         if not Path(self._state.settings.data.game_path, "D2R.exe").exists():
             raise ProcessingError(
                 f"Could not find 'D2R.exe' in '{self._state.settings.data.game_path}'"
             )
 
-        if account.auth_method == AuthMethod.Token:
-            raise ProcessingError(
-                "Token-based authentication is not supported under Linux/Wine."
-            )
+        pid: int | None = self.umu_manager.start(account)
 
-        if (
-            len(account.email) == 0
-            or account.password is None
-            or len(account.password) == 0
-        ):
-            raise ProcessingError("No email/password provided")
+        if pid and self._is_d2r_started(pid):
+            self._rename_window_title(pid, account)
+
+        return None, account, pid
 
     def _is_d2r_started(self, parent_pid: int):
         parent = psutil.Process(parent_pid)
@@ -168,6 +135,9 @@ class ProcessManager(QObject):
         #
         # For now, with XWayland we can use xdotool or wmctrl to change
         # the window title
+        #
+        # Idea: What might work is, to rename the Window in Wine
+        #       using SetWindowTextA (has to be implemented in d2rreg though)
         process = self._get_d2r_exe_pid(pid)
         if process is None:
             logger.error(f"Couldn't set window title for pid {pid}. No process found.")